requirements.txt
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
Pillow==9.1.0
drf-yasg==1.20.0
django-phonenumber-field==6.1.0 
phonenumbers==8.12.47